pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiteam"
version = "0.1.0"
description = "Agile Agentic AI system with Product Owner and Agile Coach orchestrating dynamic expert agents."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Aiteam" }
]
keywords = ["agentic ai", "agile", "django", "langchain", "autogen", "neo4j", "redis", "ollama"]

# Core runtime dependencies
dependencies = [
  "Django>=5.0,<6.0",
  "djangorestframework>=3.15",
  "channels>=4.0",
  "channels-redis>=4.2",
  "daphne>=4.0",
  "redis>=5.0",
  "celery>=5.3",
  "structlog>=24.1.0",
  "pydantic>=2.6",
  "neo4j>=5.20",
  "requests>=2.32",
  "beautifulsoup4>=4.12",
  "httpx>=0.27",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
agents = [
  "langchain>=0.2.5",
  "langchain-community>=0.2.5",
  "autogen-agentchat>=0.2.0",
  "faiss-cpu>=1.8.0",
  "openai>=1.35.0",
]
web = [
  "playwright>=1.46.0",
]
docs = [
  "sphinx>=7.3",
  "myst-parser>=2.0.0",
  "sphinx-autodoc-typehints>=2.0.0",
]
dev = [
  "black>=24.4",
  "ruff>=0.5.0",
  "isort>=5.13.2",
  "pylint>=3.2.0",
  "pylint-django>=2.5.5",
  "pytest>=8.2.0",
  "pytest-django>=4.8.0",
  "pytest-cov>=5.0.0",
  "factory-boy>=3.3.0",
  "types-requests",
  "mypy>=1.10.0",
  "pre-commit>=3.7.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["aiteam*"]

[tool.setuptools.package-data]
"*" = ["*.html", "*.css", "*.js"]

[tool.black]
line-length = 100
target-version = ["py311", "py310"]
exclude = "/(\\.venv|build|dist|migrations)/"

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["migrations", "docs/_build"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "ERA"]

[tool.ruff.lint.isort]
known-first-party = ["aiteam"]
combine-as-imports = true

[tool.isort]
profile = "black"
line_length = 100
known_first_party = "aiteam"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "aiteam.settings"
pythonpath = ["."]
addopts = "-q --disable-warnings --maxfail=1"
testpaths = ["tests"]

[tool.pylint.MASTER]
load-plugins = "pylint_django"
django-settings-module = "aiteam.settings"
ignore-patterns = "migrations,docs"

[tool.pylint.'MESSAGES CONTROL']
disable = ["R0903", "W1203"]

[tool.pylint.BASIC]
good-names = ["i", "j", "k", "ex", "Run", "_"]

